requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.4
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
pytest==8.0.0
httpx==0.27.0
python-dotenv==1.0.1
joblib==1.3.2